+++ src/lib/djerba/plugins/wgts/snv_indel_tools/tests/testthat/test-supporting-functions.R
@@ -1,0 +1,4 @@
+library(testthat)
+
+##TODO: test procVEP(maf_df)
+##TODO: test preProcRNA(gepfile, enscon)